==> Chaoran/1 data observation.ipynb <==
%matplotlib inline
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.preprocessing import LabelEncoder
from sklearn.linear_model import LinearRegression,LogisticRegression
from sklearn.model_selection import GridSearchCV

df=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\diabetic_data.csv")

# First focused on the column 24 to column 48. Those are the columns with the records of 25 types of medicines
original_df=df.copy()
original_df.T.iloc[24:49,]

#Count for each value 
med_col=original_df.iloc[:,24:47].apply(pd.value_counts).T
med_col

#check the percentage of non-No values of each medicine.
# For columns has <0.1% of data, do we need to consider a drop?
temp=round(med_col[['Down','Steady','Up']].T.sum()/med_col.T.sum(),4)
sorted(zip(med_col.index,temp),key=lambda x:x[1],reverse=True)

# Does "No" mean not taking this kind of medicine or no impact after taking the medicine?
* acetohexamide -  has only one record
* chlorpropamide - has only 86 records
* tolbutamide - has only 23 records
* Similar situation to examide ,citoglipton, glimepiride-pioglitazone,metformin-rosiglitazone,metformin-pioglitazone

for name in original_df.columns:
    print (original_df[name].value_counts(),'\n')

# Race - 2273 ? value 
# Gender - 3 unknow/Invalid value
# Weight - 98569 ? value
# Payer_code - 40256 ? value
# medical_specialty - 49949 ? value
# diag_1 - 21 ? value
# diag_2 - 358 ? value
# diag_3 - 1423 ? value

original_df.columns

original_df[original_df['diag_1']=="?"]
