==> Chaoran/3 data cleaning based on the split test_train set.ipynb <==
## The original datasets I have used are the two train_multi and test_multi csvs provided by Devon 

%matplotlib inline
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.preprocessing import LabelEncoder
from sklearn.linear_model import LinearRegression,LogisticRegression
from sklearn.model_selection import GridSearchCV

df1=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\train_multi.csv")
df2=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\test_multi.csv")

train_df=df1.copy()
test_df=df2.copy()

for col in test_df:
    series=test_df[col]
    series[series.isna()]="unknown"
for col in train_df:
    series=train_df[col]
    series[series.isna()]="unknown"

### 1 . Remove weight column

train_df.drop('weight',axis=1, inplace=True)
test_df.drop('weight',axis=1,inplace=True)

print (train_df.shape,test_df.shape)

### 2. Remove 3 rows that "gender"="Unknow/Invalid value"

train_df.drop(train_df[train_df['gender']=="Unknown/Invalid"].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['gender']=="Unknown/Invalid"].index,axis=0,inplace=True)

print (train_df.shape,test_df.shape)

### 3. Drop rows that "race"="unknown"

train_df.drop(train_df[train_df['race']=="unknown"].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['race']=="unknown"].index,axis=0,inplace=True)

print (train_df.shape,test_df.shape)

### 4. Drop column "payer_code"

train_df.drop('payer_code',axis=1, inplace=True)
test_df.drop('payer_code',axis=1, inplace=True)

print (train_df.shape,test_df.shape)

### 5. remove rows that are unknown in diag_1, diag_2 and diag_3
* originally diag_1 has 21 unknwon 
* diag_2 has 358 unknown
* diag_3 has 1423 unknown

Some rows may have both diag_1 and diag_2 as unknown , also some rows with unknown values got removed in the first 4 steps. 
Therefore, in total in this step we have 1440 rows removed

train_df.drop(train_df[train_df['diag_1']=="unknown"].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['diag_1']=="unknown"].index,axis=0,inplace=True)

train_df.drop(train_df[train_df['diag_2']=="unknown"].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['diag_2']=="unknown"].index,axis=0,inplace=True)

train_df.drop(train_df[train_df['diag_3']=="unknown"].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['diag_3']=="unknown"].index,axis=0,inplace=True)

print (train_df.shape,test_df.shape)

### 6. remove the following columns because of too few real data
* 'tolbutamide', 23 records
* 'glipizide-metformin', 13 records
* 'acetohexamide', 1 record
* 'troglitazone', 3 record
* 'examide', 0 record
* 'citoglipton', 0 record
* 'glimepiride-pioglitazone', 1 record
* 'metformin-rosiglitazone', 2 records
* 'metformin-pioglitazone', 1 record

train_df.drop(['tolbutamide','glipizide.metformin','acetohexamide','troglitazone','examide',
                'citoglipton','glimepiride.pioglitazone','metformin.rosiglitazone','metformin.pioglitazone'],axis=1, inplace=True)

test_df.drop(['tolbutamide','glipizide.metformin','acetohexamide','troglitazone','examide',
                'citoglipton','glimepiride.pioglitazone','metformin.rosiglitazone','metformin.pioglitazone'],axis=1, inplace=True)

print (train_df.shape,test_df.shape)

### 7. Based on our discussion , remove "discharge_disposition_id"= "expired" rows (id = 11,19,20,21)
* 11 - 1606 rows
* 19 - 8 rows
* 20 - 2 rows
* 21 - 0 rows

train_df.drop(train_df[train_df['discharge_disposition_id']==11].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['discharge_disposition_id']==11].index,axis=0,inplace=True)

train_df.drop(train_df[train_df['discharge_disposition_id']==19].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['discharge_disposition_id']==19].index,axis=0,inplace=True)

train_df.drop(train_df[train_df['discharge_disposition_id']==20].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['discharge_disposition_id']==20].index,axis=0,inplace=True)

print (train_df.shape,test_df.shape)

### 8. Based on the paper, remove " discharge_disposition_id"="discharge to hospice" (id=13,14)
* 13-396 rows
* 14-368 rows

train_df.drop(train_df[train_df['discharge_disposition_id']==13].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['discharge_disposition_id']==13].index,axis=0,inplace=True)

train_df.drop(train_df[train_df['discharge_disposition_id']==14].index,axis=0,inplace=True)
test_df.drop(test_df[test_df['discharge_disposition_id']==14].index,axis=0,inplace=True)

print (train_df.shape,test_df.shape)

### 9. drop "encounter_id" and "patient_nbr" column
* in my understanding these are just  unique IDs for the patients

train_df.drop(['encounter_id','patient_nbr'],axis=1, inplace=True)
test_df.drop(['encounter_id','patient_nbr'],axis=1, inplace=True)

print (train_df.shape,test_df.shape)

### 10. update "readmitted" column to "yes" or "no"
* "<30" replace with "yes"
* all others replace with "no"

train_df['readmitted']=train_df['readmitted'].replace(["<30",">30"],["Yes","NO"])
test_df['readmitted']=test_df['readmitted'].replace(["<30",">30"],["Yes","NO"])

set(test_df['readmitted'])

### 11. We have around 800 different categories in each diag_ column, combine them to the 9 main categories based on the paper

len(set(train_df['diag_1']))

# The replace_diag function can be used to replace the ids in the diag_1,diag_2 and diag_3 to the 9 main categories
def replace_diag(df):
    Circulatory_list = [str(i) for i in range(350,460)]
    Respiratory_list = [str(i) for i in range(460,520)]
    Digestive_list = [str(i) for i in range(520,589)]    
    Injury_list = [str(i) for i in range(800,1000)]
    Musculoskeletal_list=[str(i) for i in range(710,740)]
    Genitourinary_list=[str(i) for i in range(589,630)]
    Neoplasms_list=[str(i) for i in range(140,240)]
    
    for i in df:
        if  (i in Circulatory_list) or i=='785':
            df.replace(i,'Circulatory',inplace=True)
        elif (i in Respiratory_list) or i =='786':
            df.replace(i,'Respiratory', inplace=True)
        elif (i in Digestive_list) or i =='787':
            df.replace(i,'Digestive', inplace=True)
        elif (i in Injury_list):
            df.replace(i,'Injury', inplace=True)
        elif (i in Musculoskeletal_list):
            df.replace(i,'Musculoskeletal', inplace=True)
        elif (i in Genitourinary_list) or i =='788':
            df.replace(i,'Genitourinary', inplace=True)
        elif (i in Neoplasms_list):
            df.replace(i,'Neoplasms', inplace=True)
        elif len(i)>3:
            df.replace(i,'Diabetes', inplace=True)
        else:
            df.replace(i,'Other',inplace=True)
            

replace_diag(train_df['diag_1'])
replace_diag(test_df['diag_1'])
print (len(set(train_df['diag_1'])),len(set(test_df['diag_1'])))

replace_diag(train_df['diag_2'])
replace_diag(test_df['diag_2'])
print (len(set(train_df['diag_2'])),len(set(test_df['diag_2'])))

replace_diag(train_df['diag_3'])
replace_diag(test_df['diag_3'])
print (len(set(train_df['diag_3'])),len(set(test_df['diag_3'])))

### Okay, we get our origin dataset, save it as "cleantrain.csv" and "cleantest.csv" first. Then we move to encode the nominal columns 

print (train_df.shape,test_df.shape)

train_df.to_csv("cleantrain.csv")
test_df.to_csv("cleantest.csv")

* **


### 12. Encode the following columns for xgboost 
race,gender,age,medical_specialty,max_glu_serum,A1Cresult,metformin,repaglinide,nateglinide,
chlorpropamide,glimepiride,glipizide,glyburide,pioglitazone,rosiglitazone,acarbose,miglitol,tolazamide,insulin,
glyburide-metformin,change,diabetesMed

#copy origin_df to encoder_df, so we don't make changes on origin_df
train=pd.read_csv('cleantrain.csv',index_col=0)
test=pd.read_csv('cleantest.csv',index_col=0)
lencoder=LabelEncoder()
ind="_labels"
encoder_list=['race','gender','age','medical_specialty','max_glu_serum','A1Cresult','metformin','repaglinide','nateglinide',
        'chlorpropamide','glimepiride','glipizide','glyburide','pioglitazone','rosiglitazone','acarbose','miglitol',
        'tolazamide','insulin','glyburide.metformin','change','diabetesMed','diag_1', 'diag_2', 'diag_3']

numeric_col=['admission_type_id','discharge_disposition_id','admission_source_id','time_in_hospital','num_lab_procedures',
            'num_procedures','num_medications','number_outpatient','number_emergency','number_inpatient','number_diagnoses',
            'readmitted']

train.columns

for i in encoder_list:
    lencoder.fit(train[i])
    test[i] = test[i].map(lambda s: 'unknown' if s not in lencoder.classes_ else s)
    lencoder.classes_ = np.append(lencoder.classes_, 'unknown')
    
    train[i]=lencoder.transform(train[i]).reshape(-1,1)
    test[i]=lencoder.transform(test[i]).reshape(-1,1)

train['readmitted']=train['readmitted'].replace(["Yes","NO"],[1,0])
train.head(50)

test['readmitted']=test['readmitted'].replace(["Yes","NO"],[1,0])

# Save the label encoded dfs to csv files
train.to_csv('le_train.csv')
test.to_csv('le_test.csv')
