==> Chaoran/4 use RandomSearchCV to look for best parameters for xgb.ipynb <==
%matplotlib inline
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.ensemble import GradientBoostingRegressor,GradientBoostingClassifier
from sklearn.preprocessing import LabelEncoder
from sklearn.linear_model import LinearRegression,LogisticRegression
from sklearn.model_selection import GridSearchCV,RandomizedSearchCV
import xgboost as xgb
df=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\git\healthcare_capstone\Chaoran\le_train.csv",index_col=0)

trainX=df.drop('readmitted',axis=1).copy()
trainY=df['readmitted']

para= {"learning_rate" : [0.01,0.1,1],
       "n_estimators": [200,500,1000,2000,3000,4000],
       "max_depth" : [3,4,5,6],
       "subsample" : [0.8,0.9,1],
       "colsample_bytree" : [0.8,0.9,1]}

#XGBregressor
xgboost = xgb.XGBRegressor()
n_iter_search = 2
ran_search_xgb=RandomizedSearchCV(estimator=xgboost, param_distributions=para,n_iter=n_iter_search,
                                  scoring='neg_mean_squared_error', cv=5, return_train_score=True)


xgbmodel=ran_search_xgb.fit(trainX,trainY)

print("Best parameters found: ", xgbmodel.best_params_)
print("Lowest RMSE found: ", np.sqrt(np.abs(xgbmodel.best_score_)))

#Best parameters found:  {'subsample': 0.8, 'n_estimators': 3000, 'max_depth': 6, 'learning_rate': 0.1, 'colsample_bytree': 1}
#Lowest RMSE found:  0.3404059987207308

#XGBClassifier
xgboost2 = xgb.XGBClassifier()
n_iter_search2 =50
ran_search_xgb2=RandomizedSearchCV(estimator=xgboost, param_distributions=para,n_iter=n_iter_search2,
                                  scoring='roc_auc', cv=5, return_train_score=True)

import datetime
starttime = datetime.datetime.now()

xgbmodel2=ran_search_xgb2.fit(trainX,trainY)

endtime=datetime.datetime.now()
print ((endtime-starttime).seconds)

print("Best parameters found: ", xgbmodel2.best_params_)
print("Lowest RMSE found: ", np.sqrt(np.abs(xgbmodel2.best_score_)))
