==> Chaoran/6 Xgboost with top 10 important features.ipynb <==
%matplotlib inline
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.ensemble import GradientBoostingRegressor,GradientBoostingClassifier
from sklearn.preprocessing import LabelEncoder
from sklearn.linear_model import LinearRegression,LogisticRegression
from sklearn.model_selection import GridSearchCV,RandomizedSearchCV
import xgboost as xgb
from sklearn.metrics import roc_auc_score
from sklearn.metrics import accuracy_score

df=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\git\healthcare_capstone\Chaoran\le_train.csv",index_col=0)
trainX=df[['num_lab_procedures','num_medications','time_in_hospital','age','medical_specialty',
          'discharge_disposition_id','diag_1','number_inpatient','diag_2','diag_3']].copy()
trainY=df['readmitted']

df2=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\git\healthcare_capstone\Chaoran\le_test.csv",index_col=0)
testX=df2[['num_lab_procedures','num_medications','time_in_hospital','age','medical_specialty',
          'discharge_disposition_id','diag_1','number_inpatient','diag_2','diag_3']].copy()
testY=df2['readmitted']

### Regressor

xgbr= xgb.XGBRegressor()
xgbr.set_params(learning_rate=0.1,subsample=0.8,n_estimators=3000,max_depth=6,
                colsample_bytree=1)

xgbr.fit(trainX,trainY)

train_predict=xgbr.predict(trainX)
train_prediction=[round(value) for value in train_predict]
train_accuracy=accuracy_score(trainY,train_prediction)
print ("train accurary:", train_accuracy)

test_predict=xgbr.predict(testX)
test_prediction=[round(value) for value in test_predict]
test_accuracy=accuracy_score(testY,test_prediction)
print("test accurary:" , test_accuracy)

### Classifier

xgbc= xgb.XGBClassifier()
xgbc.set_params(learning_rate=0.1,subsample=0.8,n_estimators=3000,max_depth=6,
                colsample_bytree=1,objective='binary:logistic')

xgbc.fit(trainX,trainY)

train_predict2=xgbc.predict(trainX)

test_predict2=xgbc.predict(testX)

print ("train auc: ",roc_auc_score(trainY,train_predict2))
print ("test auc: ",roc_auc_score(testY,test_predict2))

# Try 20 features

df3=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\git\healthcare_capstone\Chaoran\le_train.csv",index_col=0)
trainX=df3[['num_lab_procedures','num_medications','time_in_hospital','age','medical_specialty',
          'discharge_disposition_id','diag_1','number_inpatient','diag_2','diag_3','num_procedures',
           'number_diagnoses','admission_type_id','insulin','number_outpatient','race','admission_source_id',
           'number_emergency','A1Cresult','gender']].copy()
trainY=df3['readmitted']

df4=pd.read_csv(r"C:\nydsa bootcamp slides\Projects\4\git\healthcare_capstone\Chaoran\le_test.csv",index_col=0)
testX=df4[['num_lab_procedures','num_medications','time_in_hospital','age','medical_specialty',
          'discharge_disposition_id','diag_1','number_inpatient','diag_2','diag_3','num_procedures',
           'number_diagnoses','admission_type_id','insulin','number_outpatient','race','admission_source_id',
           'number_emergency','A1Cresult','gender']].copy()
testY=df4['readmitted']

xgbc.fit(trainX,trainY)

train_predict2=xgbc.predict(trainX)

test_predict2=xgbc.predict(testX)

print ("train auc: ",roc_auc_score(trainY,train_predict2))
print ("test auc: ",roc_auc_score(testY,test_predict2))
